clean architecture mod
