+++ back-end/helpers/post/post_transcript_interactor.js
@@ -1,9 +1,9 @@
-import {InputBoundaryInterface} from "../../interfaces/input-boundary-interface.js";
+import { FormatTranscriptInterface } from "../../interfaces/format-transcript-interface.js";
 import TranscriptFormatter from "./transcript_data_formatter.js";
 import OutputDataBoundary from "../general/output_data_boundary.js";
 import { IntentInterface } from "../../interfaces/intent-interface.js";
 
-export default class PostTranscriptInteractor extends InputBoundaryInterface{
+export default class PostTranscriptInteractor extends FormatTranscriptInterface{
     static #IntentDao;
 
     static setIntentDao(dao){
