+++ back-end/interfaces/input-boundary-interface.js
@@ -2,39 +2,15 @@
 /**
  * Abstract Class InputBoundaryInterface.
  *
- * @class Interface
+ * @class Input boundary
+ * Inherited by all other input interfaces
  */
 export class InputBoundaryInterface{
     static isInputBoundaryInterface = true;
-
     /**
-     * @param {Object} query find specific intent(s)
+     * @param {Object} dao data access object
      */
-    static async getTranscript(query = {}){
-        throw new Error("not implemented");
-    }
-
-    /**
-     *  @param {Object} rawTranscript raw transcript uploaded by user
-     *  @return {Object} properly formatted data from the transcript
-     * */
-    static async formatTranscript(rawTranscript = {}) {
-        throw new Error("not implemented");
-    }
-
-    /**
-     *  @param {Object} filter object passed to mangoose to find the question
-     *  @param {Object} status the star status
-     * */
-     static async setStarStatus(filter = {}, status = {}) {
-        throw new Error("not implemented");
-    }
-
-    /**
-     *  @param {Object} filter object passed to mangoose to find the question
-     *  @param {Object} status the flag status
-     * */
-     static async setFlagStatus(filter = {}, status = {}) {
+     static async setIntentDao(dao){
         throw new Error("not implemented");
     }
 }
